Added DbC class for Block. Changed offsets to use size_t instead of unsigned in order to support objects larger than 4GB

+++ src/api/cuda/Accelerator.h
@@ -143,8 +143,8 @@ public:
     TESTABLE gmacError_t copyAccelerator(accptr_t dst, const accptr_t src, size_t size);
 
     /* Asynchronous interface */
-    TESTABLE gmacError_t copyToAcceleratorAsync(accptr_t acc, IOBuffer &buffer, unsigned bufferOff, size_t count, Mode &mode, CUstream stream);
-    TESTABLE gmacError_t copyToHostAsync(IOBuffer &buffer, unsigned bufferOff, const accptr_t acc, size_t count, Mode &mode, CUstream stream);
+    TESTABLE gmacError_t copyToAcceleratorAsync(accptr_t acc, IOBuffer &buffer, size_t bufferOff, size_t count, Mode &mode, CUstream stream);
+    TESTABLE gmacError_t copyToHostAsync(IOBuffer &buffer, size_t bufferOff, const accptr_t acc, size_t count, Mode &mode, CUstream stream);
     CUstream createCUstream();
     void destroyCUstream(CUstream stream);
     CUresult queryCUstream(CUstream stream);
