+++ src/gmac/stdc/io.cpp
@@ -44,7 +44,7 @@ size_t SYMBOL(fread)(void *buf, size_t size, size_t nmemb, FILE *stream)
     size_t n = size * nmemb;
     size_t ret = 0;
 
-    unsigned off = 0;
+    size_t off = 0;
     size_t bufferSize = paramPageSize > size ? paramPageSize : size;
     Mode &mode = Mode::current();
     IOBuffer *buffer = mode.createIOBuffer(bufferSize);
@@ -62,8 +62,8 @@ size_t SYMBOL(fread)(void *buf, size_t size, size_t nmemb, FILE *stream)
         err = buffer->wait();
         ASSERTION(err == gmacSuccess);
 
-        left -= (size * elems);
-        off  += unsigned(size * elems);
+        left -= size * elems;
+        off  += size * elems;
         TRACE(GLOBAL, FMT_SIZE" of %zd bytes read", elems * size, nmemb * size);
     }
     mode.destroyIOBuffer(buffer);
@@ -94,7 +94,7 @@ size_t SYMBOL(fwrite)(const void *buf, size_t size, size_t nmemb, FILE *stream)
     size_t n = size * nmemb;
     size_t ret = 0;
 
-    unsigned off = 0;
+    size_t off = 0;
     size_t bufferSize = paramPageSize > size ? paramPageSize : size;
     Mode &mode = Mode::current();
     IOBuffer *buffer = mode.createIOBuffer(bufferSize);
@@ -114,7 +114,7 @@ size_t SYMBOL(fwrite)(const void *buf, size_t size, size_t nmemb, FILE *stream)
         ret += elems;
         
         left -= size * elems;
-        off  += unsigned(size * elems);
+        off  += size * elems;
 
         TRACE(GLOBAL, FMT_SIZE" of "FMT_SIZE" bytes written", elems * size, nmemb * size);
     }
